Add files via upload
In this capstone, we will predict if the Falcon 9 first stage will land successfully. SpaceX advertises Falcon 9 rocket launches on its website with a cost of 62 million dollars; other providers cost upward of 165 million dollars each, much of the savings is because SpaceX can reuse the first stage. Therefore if we can determine if the first stage will land, we can determine the cost of a launch. This information can be used if an alternate company wants to bid against SpaceX for a rocket launch. In this lab, we collect and make sure the data is in the correct format from an API.
In this lab, we will make a get request to the SpaceX API. We will also do some basic data wrangling and formating.
-Request to the SpaceX API
-Clean the requested data